// lib/lead-import-types.ts
export type ImportedLeadDraft = {
  firstName: string;
  lastName: string;
  ownerName: string;
  phone: string;
  email: string;
  propertyAddress: string;
  city: string;
  state: string;
  zipCode: string;
  mailingAddress: string;
  county: string;
  parcelId: string;
  situationDetails: string;
  source: string;
};